+++ ui-v2/lib/commands/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "commands",
+  "keywords": [
+    "ember-addon"
+  ]
+}
