+++ ui-v2/package.json
@@ -38,11 +38,14 @@
     ]
   },
   "devDependencies": {
+    "@babel/core": "^7.2.2",
+    "@babel/node": "^7.2.2",
     "@hashicorp/consul-api-double": "^2.0.1",
     "@hashicorp/ember-cli-api-double": "^1.3.0",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "base64-js": "^1.3.0",
     "broccoli-asset-rev": "^2.4.5",
+    "chalk": "^2.4.2",
     "dart-sass": "^1.14.1",
     "ember-ajax": "^3.0.0",
     "ember-block-slots": "^1.1.11",
@@ -104,6 +107,7 @@
   },
   "ember-addon": {
     "paths": [
+      "lib/commands",
       "lib/startup"
     ]
   }
