+++ ui-v2/tests/steps/debug/index.js
@@ -1,0 +1,20 @@
+/* eslint no-console: "off" */
+export default function(scenario, assert, currentURL) {
+  scenario
+    .then('print the current url', function(url) {
+      console.log(currentURL());
+      return Promise.resolve();
+    })
+    .then('log the "$text"', function(text) {
+      console.log(text);
+      return Promise.resolve();
+    })
+    .then('pause for $milliseconds', function(milliseconds) {
+      return new Promise(function(resolve) {
+        setTimeout(resolve, milliseconds);
+      });
+    })
+    .then('ok', function() {
+      assert.ok(true);
+    });
+}
